pyproject.toml
[tool.poetry]
name = "dartboi"
version = "0.1.0"
description = ""
authors = ["Samuel Čavoj <samuel@cavoj.net>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^12.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
